[issue #28] check $ref on create
If a object was created or updated (which is essentially the same) the
file system core did not check if any $ref pointer was actually present.
This can cause the data to become inconsistent which then leads to other
issues because the implementation relies on data consistency. Now the
implementation checks if a reference is valid (pointing to an object
already present) or is about to be created by the current write
activity.