new jcommander groupid and no active default profiles
The currently used jcommander version is a cismet fork. Thus the groupid
must be one of the cismet realm to avoid clashes with the original
version. Additionally, the jcommander dependency has to be explicitely
excluded from the testng dependency as it does make use of another
version of jcommander which causes issues for unit tests.

The case-insensitive-test-override profile has been activated by default
accidently in some feature branch. This is reverted because it is only a
helper profile for partitions that are not formatted with a
case-sensitive FS.