fixed #60
classKey now does not contain the domainname anymore.
The EntitiesAPI has been changed to pass the classKey to the EntityCore
without the domainname, according to the new definition.

The FileSystenEntityCore has been changed, so that it intreprets
the classKey the new way. This affects mainly the buildRef method,
that now longer splits the given classKey to obtain the className
and the domainName. Instead, the classKey represents the className,
and the domainName is retrieved from the current Server instance.

The test classes has been modified, so that they match the new
classKey definition.