fixed #64
New parameter "requestInstanceType" added to the "createNewActionTask" method of the action API.
New method "executeNewAction" in the ActionCore interface.

If the value of "requestInstanceType" is "task" (default), the "createNewActionTask" core method is invoked.
If the value is "result", the new "executeNewAction" method is invoked.

The "requestResultingInstance" parameter of "createNewActionTask" (ActionCore) is set to deprecated. Reason:
This method is always fast and has a small output (the action task without the result). There is no need
to suppress the result.